App/API/requirements.txt
Django
djangorestframework
pandas
psycopg2
django-cors-headers
requests